requirements.txt
attrs==20.3.0
biopython==1.78
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-daq==0.5.0
dash-extensions==0.0.45
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
eli5==0.11.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.9.0
fontawesome==5.10.1.post1
future==0.18.2
graphviz==0.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
kiwisolver==1.3.0
MarkupSafe==1.1.1
more-itertools==8.7.0
networkx==2.5
numpy==1.20.1
pandas==1.2.2
Pillow==8.0.1
plotly==4.11.0
pyhull==2015.2.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sortedcontainers==2.2.2
tabulate==0.8.8
threadpoolctl==2.1.0
tifffile==2020.9.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
